backend/pyproject.toml
[tool.poetry]
name = "mykolaiv_city_bot_backend"
version = "0.1.0"
description = ""
authors = ["cheerfulleg <yolginivan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
python-dotenv = "^0.20.0"
uvicorn = "^0.17.6"
motor = "^2.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

